Added missing packages and made example test
