+++ src/lib/domain/student/index.ts
@@ -1,4 +1,5 @@
 import * as student_interfaces from './interfaces';
+import * as entities from './entities'
 import StudentService from './services/student_service';
 import StudentRepository from './repositories/student_repository';
 import UniversityRepository from './repositories/university_repository';
@@ -10,6 +11,7 @@ import {S_TYPES} from './types'
 export {
     S_TYPES,
     student_interfaces,
+    entities,
     StudentService,
     StudentRepository,
     StudyProgramRepository,
