--- src/lib/test/student_service.test.ts
@@ -1,11 +1,0 @@
-import container from '../container';
-import * as st from '../domain/student'
-
-
-
-const ser = container.resolve(st.StudentService)
-
-
-test(" Dimelo ", async ( ) => {
-    
-})
